feat: add lowess

+++ src/knn.rs
@@ -28,7 +28,7 @@ pub struct Knn<M: DistanceMetric<f64, DIMENSIONS>> {
     window: WindowType,
     kd_tree: KdTree<f64, usize, DIMENSIONS, BUCKET_SIZE, u32>,
     data: Vec<Data>,
-    weights: Vec<f32>,
+    weights: Vec<f64>,
     _marker: PhantomData<M>,
 }
 
@@ -52,7 +52,7 @@ impl<M: DistanceMetric<f64, DIMENSIONS>> Knn<M> {
         }
     }
 
-    pub fn fit(&mut self, data: Vec<Data>, weights: Option<Vec<f32>>) {
+    pub fn fit(&mut self, data: Vec<Data>, weights: Option<Vec<f64>>) {
         self.data = data;
         self.weights = weights.unwrap_or_else(|| vec![1.0; self.data.len()]);
 
@@ -75,12 +75,12 @@ impl<M: DistanceMetric<f64, DIMENSIONS>> Knn<M> {
     fn predict_class(
         kernel_distances: &[f64],
         targets: &[Diagnosis],
-        weights: &[f32],
+        weights: &[f64],
     ) -> Diagnosis {
         let mut class_scores: HashMap<Diagnosis, f64> = HashMap::new();
 
         for (i, target) in targets.iter().enumerate() {
-            let weighted_score = kernel_distances[i] * weights[i] as f64;
+            let weighted_score = kernel_distances[i] * weights[i];
             *class_scores.entry(*target).or_insert(0.0) += weighted_score;
         }
 
@@ -94,7 +94,7 @@ impl<M: DistanceMetric<f64, DIMENSIONS>> Knn<M> {
     fn predict_with_neighbors(
         &self,
         x: &[f64; DIMENSIONS],
-    ) -> (Vec<f64>, Vec<Diagnosis>, Vec<f32>) {
+    ) -> (Vec<f64>, Vec<Diagnosis>, Vec<f64>) {
         let (distances, indices): (Vec<f64>, Vec<usize>) = match self.window {
             WindowType::Fixed => self.kd_tree.within::<M>(x, self.radius.powi(2)),
             WindowType::Unfixed => self.kd_tree.nearest_n::<M>(x, self.k),
